\name{biertest.dat}
\alias{biertest.dat}
\docType{data}
\title{The Beer Testing Data in Kunert (1998)}
\description{
  The data comes from a beer testing experiment where assessors had to rate
  the bitterness of five beers.    
}
\usage{data(biertest.dat)}
\format{
A matrix with 12 rows corresponding to the assessors, 5 columns corresponding to 
  the five periods.
}
\details{
  The possible values for the ratings lie in the interval [0, 12].
  Note that the data were reordered after the experiment. }
\source{
  Kunert, J. (1998): Sensory experiments as crossover studies. Food Quality and Preference 9, 243-253
  (table 1). 
}

\examples{
data(biertest.dat)
}
\keyword{datasets}
