\name{biertest.f2}
\alias{biertest.f2}
\docType{data}
\title{Experimental Design for the Beer Testing Data in Kunert (1998)}
\description{
  The design is a row-column design. It is actually a generalized latin square design that is also carryover balanced. 
}
\usage{data(biertest.d4)}
\format{
A matrix with 10 rows corresponding to the first 10 assessors in a fictional uniformity trial, 
  5 columns corresponding to the five periods. 
}
\details{
  There are five products to be tested, numbered 1,\ldots,5.
}
\source{
  Kunert, J. (1998): Sensory experiments as crossover studies. Food Quality and Preference 9, 243-253
  (design f2). 
}

\examples{
data(biertest.f2)
}
\keyword{datasets} % at least one, from doc/KEYWORDS
