\name{crossdes-internal}
\alias{choices}
\alias{design.row}
\alias{factor.comb}
\alias{GF}
\alias{mult}
\alias{Q.t}
\alias{random.AZ}
\alias{random.RT}
\alias{redu}
\alias{redu.modp}
\alias{tcdf.plot}
\alias{Td}
\title{ Internal Functions in crossdes }
\description{
  For internal use only.}
\author{ Oliver Sailer \email{sailer@statistik.uni-dortmund.de}}

\keyword{internal}
