\name{crossdes-internal}
\alias{choices}
\alias{design.row}
\alias{factor.comb}
\alias{GF}
\alias{mult}
\alias{Q.t}
\alias{random.AZ}
\alias{random.azais}
\alias{random.bailey}
\alias{random.RC}
\alias{redu}
\alias{redu.modp}
\alias{tcdf.plot}
\alias{Td}
\title{ Internal Functions in crossdes }
\description{
  For internal use only.}
\usage{
design.row(d)
}
\author{ Oliver Sailer \email{sailer@statistik.uni-dortmund.de}}

\note{ The function \code{design.row} requires the library \code{gtools}. 
}

\keyword{internal}
